lintian fixes for windows

+++ gcl/h/att.h
@@ -72,7 +72,7 @@
 
 #define INIT_ALLOC \
      	heap_end = sbrk(0); \
-	({fixnum i;if ((i = ((int)heap_end & (PAGESIZE - 1))))	\
+	({fixnum i;if ((i = ((unsigned long)heap_end & (PAGESIZE - 1))))	\
 		     sbrk(PAGESIZE - i);});			\
 	heap_end = core_end = sbrk(0);
